Changing the calculaPrecoUnitario() and calculaPrecoUnitarioTest() code
 - Test passing

+++ src/tesouro/LTN.java
@@ -4,7 +4,7 @@ public class LTN {
 
 	public double calculaPrecoUnitario(double taxaIndicativa) {
 			
-		double denominador = Math.pow((1 + taxaIndicativa), 1/252);
+		double denominador = Math.pow(1 + taxaIndicativa, (double)1/252);
 		double precoUnitario = 1000/denominador;
 		
 		return precoUnitario;
